日常更新

+++ src/main/java/com/anxu/smarthomeunity/model/dto/wecommunity/ChatHistoryQueryDTO.java
@@ -12,7 +12,7 @@ public class ChatHistoryQueryDTO {
     // 社区ID
     private Integer communityId;
     // 锚点msgId（首次传0，后续传当前加载的最老的msgId）
-    private Long lastMsgId;
+    private Integer lastMsgId;
     // 每页条数（固定50，也可前端自定义）
     private Integer pageSize = 50;
 }
